+++ src/containers/Chat/ChatMessages/ChatMessages.test.tsx
@@ -1,24 +1,27 @@
-import { render, act, screen } from '@testing-library/react';
 import 'mocks/matchMediaMock';
-import { ApolloClient, ApolloProvider, InMemoryCache } from '@apollo/client';
-import { fireEvent, waitFor } from '@testing-library/dom';
+import { InMemoryCache } from '@apollo/client';
+import { MockedProvider } from '@apollo/client/testing';
+import { act, fireEvent, render, screen, waitFor } from '@testing-library/react';
+import { DEFAULT_ENTITY_LIMIT, DEFAULT_MESSAGE_LIMIT } from 'common/constants';
+import { SEARCH_QUERY } from 'graphql/queries/Search';
+import ChatMessages from './ChatMessages';
 import { MemoryRouter, Route, Routes } from 'react-router';
 import { BrowserRouter as Router } from 'react-router-dom';
-import { MockedProvider } from '@apollo/client/testing';
-import { vi } from 'vitest';
-
-import { ChatMessages } from './ChatMessages';
-import { SEARCH_QUERY } from '../../../graphql/queries/Search';
-import { DEFAULT_ENTITY_LIMIT, DEFAULT_MESSAGE_LIMIT } from '../../../common/constants';
+import { getAttachmentPermissionMock } from 'mocks/Attachment';
+import { clearMessagesQuery, contactCollectionsQuery } from 'mocks/Contact';
+import { OrganizationStateMock } from 'mocks/Organization';
 import {
   CONVERSATION_MOCKS,
-  loadMoreChats,
+  createAndSendMessageMutation2,
   mocksWithConversation,
+  sendMessageInWaGroup,
+  sendMessageInWaGroupCollection,
   sendMessageMock,
-} from '../../../mocks/Chat';
-import { waGroup, waGroupcollection } from 'mocks/Groups';
+} from 'mocks/Chat';
+import { getCollectionInfo } from 'mocks/Collection';
 import { userEvent } from '@testing-library/user-event';
 import { setNotification } from 'common/notification';
+import { waGroup, waGroupcollection } from 'mocks/Groups';
 
 vi.mock('common/notification', async (importOriginal) => {
   const mod = await importOriginal<typeof import('common/notification')>();
@@ -30,50 +33,78 @@ vi.mock('common/notification', async (importOriginal) => {
   };
 });
 
-const defineUrl = (url: string) => {
-  Object.defineProperty(window, 'location', {
-    value: {
-      href: url,
-    },
-    writable: true,
-  });
-};
-
-const body = {
-  id: '1',
-  body: 'Hey there whats up?',
-  insertedAt: '2020-06-25T13:36:43Z',
-  location: null,
-  messageNumber: 48,
-  receiver: {
-    id: '1',
-  },
-  sender: {
-    id: '2',
-  },
-  type: 'TEXT',
-  media: null,
-  errors: '{}',
-  contextMessage: {
-    body: 'All good',
-    contextId: 1,
-    messageNumber: 10,
-    errors: '{}',
-    media: null,
-    type: 'TEXT',
-    insertedAt: '2021-04-26T06:13:03.832721Z',
+const messages = (limit: number, skip: number) =>
+  new Array(limit).fill(null).map((val: any, index: number) => ({
+    id: `${index + skip}`,
+    body: 'Hey there whats up?',
+    insertedAt: '2020-06-25T13:36:43Z',
     location: null,
+    messageNumber: index + skip + 4,
     receiver: {
       id: '1',
     },
     sender: {
       id: '2',
-      name: 'User',
+    },
+    type: 'TEXT',
+    media: null,
+    errors: '{}',
+    contextMessage:
+      index % 5 === 0
+        ? {
+            body: 'All good',
+            contextId: 1,
+            messageNumber: 10,
+            errors: '{}',
+            media: null,
+            type: 'TEXT',
+            insertedAt: '2021-04-26T06:13:03.832721Z',
+            location: null,
+            receiver: {
+              id: '1',
+            },
+            sender: {
+              id: '2',
+              name: 'User',
+            },
+          }
+        : null,
+    interactiveContent: '{}',
+    sendBy: 'test',
+    flowLabel: null,
+  }));
+
+const loadMoreQuery = {
+  request: {
+    query: SEARCH_QUERY,
+    variables: {
+      contactOpts: { limit: 1 },
+      filter: { id: '2' },
+      messageOpts: { limit: 22, offset: 1 },
     },
   },
-  interactiveContent: '{}',
-  sendBy: 'test',
-  flowLabel: null,
+  result: {
+    data: {
+      search: [
+        {
+          contact: {
+            id: '2',
+            name: 'Effie Cormier',
+            phone: '987654321',
+            maskedPhone: '98****321',
+            lastMessageAt: new Date(),
+            status: 'VALID',
+            fields: '{}',
+            bspStatus: 'SESSION_AND_HSM',
+            isOrgRead: true,
+          },
+          group: null,
+          id: 'contact_2',
+          messages: messages(20, 20),
+        },
+      ],
+    },
+  },
 };
 
 const cache = new InMemoryCache({ addTypename: false });
@@ -100,7 +131,7 @@ export const searchQuery = {
           bspStatus: 'SESSION_AND_HSM',
           isOrgRead: true,
         },
-        messages: [body],
+        messages: messages(20, 1),
       },
     ],
   },
@@ -131,7 +162,7 @@ export const contact = {
           bspStatus: 'SESSION_AND_HSM',
           isOrgRead: true,
         },
-        messages: [body],
+        messages: messages(10, 3),
       },
       {
         id: 'contact_1',
@@ -226,6 +257,173 @@ export const collection = {
   },
 };
 
+const coversationMock = (variables: any) => ({
+  request: {
+    query: SEARCH_QUERY,
+    variables,
+  },
+  result: {
+    data: {
+      search: [
+        {
+          id: 'contact_2',
+          group: null,
+          contact: {
+            id: '2',
+            name: 'Jane Doe',
+            phone: '9857274829',
+            maskedPhone: '974****678',
+            lastMessageAt: '2020-06-25T13:36:43Z',
+            status: 'VALID',
+            bspStatus: 'SESSION_AND_HSM',
+            isOrgRead: true,
+            fields: '{}',
+          },
+          messages: [
+            {
+              id: '1',
+              body: 'Hello',
+              insertedAt: '2020-06-25T13:36:43Z',
+              receiver: {
+                id: '1',
+              },
+              sender: {
+                id: '2',
+              },
+              type: 'TEXT',
+              media: null,
+              location: null,
+              errors: null,
+              messageNumber: 2,
+              contextMessage: {
+                body: 'All good',
+                contextId: 1,
+                messageNumber: 10,
+                errors: '{}',
+                media: null,
+                type: 'TEXT',
+                insertedAt: '2021-04-26T06:13:03.832721Z',
+                location: null,
+                receiver: {
+                  id: '1',
+                },
+                sender: {
+                  id: '2',
+                  name: 'User',
+                },
+              },
+              sendBy: 'Glific User',
+              interactiveContent: '{}',
+              flowLabel: null,
+            },
+          ],
+        },
+      ],
+    },
+  },
+});
+
+const mocks = [
+  getAttachmentPermissionMock,
+  contactCollectionsQuery(2),
+  contactCollectionsQuery(3),
+  OrganizationStateMock,
+  ...CONVERSATION_MOCKS,
+  ...mocksWithConversation,
+  coversationMock({
+    contactOpts: { limit: 1 },
+    filter: { id: '5' },
+    messageOpts: { limit: 20, offset: 0 },
+  }),
+  clearMessagesQuery,
+  loadMoreQuery,
+  getCollectionInfo({ id: '2' }),
+  getCollectionInfo({ id: '5' }),
+  coversationMock({
+    contactOpts: { limit: 1 },
+    messageOpts: { limit: 20, offset: 0 },
+    filter: { id: '2' },
+  }),
+  coversationMock({
+    contactOpts: { limit: 1 },
+    messageOpts: { limit: 20, offset: 0 },
+    filter: { id: '2', searchGroup: true },
+  }),
+  {
+    request: {
+      query: SEARCH_QUERY,
+      variables: {
+        contactOpts: { limit: 1 },
+        messageOpts: { limit: 20, offset: 0 },
+        filter: { id: '3' },
+      },
+    },
+    result: {
+      data: {
+        search: [
+          {
+            id: 'contact_3',
+            contact: {
+              id: '3',
+              name: 'New Contact',
+              phone: '9857274829',
+              maskedPhone: '974****678',
+              lastMessageAt: '2020-06-25T13:36:43Z',
+              status: 'VALID',
+              bspStatus: 'SESSION_AND_HSM',
+              isOrgRead: true,
+              fields: '{}',
+            },
+            group: null,
+            messages: [
+              {
+                id: '1',
+                body: 'Hello',
+                insertedAt: '2020-06-25T13:36:43Z',
+                receiver: {
+                  id: '1',
+                },
+                sender: {
+                  id: '3',
+                },
+                type: 'TEXT',
+                media: null,
+                location: null,
+                errors: null,
+                messageNumber: 2,
+                contextMessage: {
+                  body: 'All good',
+                  contextId: 1,
+                  messageNumber: 10,
+                  errors: '{}',
+                  media: null,
+                  type: 'TEXT',
+                  insertedAt: '2021-04-26T06:13:03.832721Z',
+                  location: null,
+                  receiver: {
+                    id: '1',
+                  },
+                  sender: {
+                    id: '3',
+                    name: 'User',
+                  },
+                },
+                sendBy: 'Glific User',
+                interactiveContent: '{}',
+                flowLabel: null,
+              },
+            ],
+          },
+        ],
+      },
+    },
+  },
+  createAndSendMessageMutation2,
+  sendMessageMock,
+  sendMessageInWaGroup,
+  sendMessageInWaGroupCollection,
+];
+
 export const collectionWithLoadMore = {
   query: SEARCH_QUERY,
   variables: {
@@ -241,26 +439,19 @@ export const collectionWithLoadMore = {
 // add collection to apollo cache
 cache.writeQuery(collection);
 
-const client = new ApolloClient({
-  cache: cache,
-  uri: 'http://localhost:4000/',
-  assumeImmutableResults: true,
-});
 window.HTMLElement.prototype.scrollIntoView = vi.fn();
-const mocks = [loadMoreChats];
 
 const chatMessages = (
   <MemoryRouter>
-    <MockedProvider mocks={mocks}>
-      <ApolloProvider client={client}>
-        <ChatMessages entityId="2" />
-      </ApolloProvider>
+    <MockedProvider mocks={mocks} cache={cache}>
+      <ChatMessages entityId="2" />
     </MockedProvider>
   </MemoryRouter>
 );
 
 it('should have title as contact name', async () => {
   const { getByTestId } = render(chatMessages);
+
   await waitFor(() => {
     expect(getByTestId('beneficiaryName')).toHaveTextContent('Effie Cormier');
   });
@@ -274,17 +465,19 @@ it('should have an emoji picker', async () => {
 });
 
 it('should contain the mock message', async () => {
-  const { getByText } = render(chatMessages);
+  const { getAllByText } = render(chatMessages);
   await waitFor(() => {
-    expect(getByText('Hey there whats up?')).toBeInTheDocument();
+    expect(getAllByText('Hey there whats up?')[0]).toBeInTheDocument();
   });
 });
 
 test('focus on the latest message', async () => {
-  const { getByTestId } = render(chatMessages);
+  const { getAllByText } = render(chatMessages);
   await waitFor(() => {
-    const message = getByTestId('message');
-    expect(message.scrollIntoView).toBeCalled();
+    const messages = getAllByText('Hey there whats up?');
+
+    // since there are 20 messages the latest is message[19]
+    expect(messages[19].scrollIntoView).toHaveBeenCalled();
   });
 });
 
@@ -303,26 +496,38 @@ test('click on Jump to latest', async () => {
   });
 });
 
-test('Contact: if not cache', async () => {
-  const chatMessagesWithCollection = (
-    <ApolloProvider client={client}>
-      <Router>
-        <MockedProvider mocks={[...CONVERSATION_MOCKS, ...mocksWithConversation]}>
-          <ChatMessages entityId="5" />
-        </MockedProvider>
-      </Router>
-    </ApolloProvider>
-  );
-  render(chatMessagesWithCollection);
+test('click on Clear conversation', async () => {
+  const { getByTestId } = render(chatMessages);
+  fireEvent.click(getByTestId('dropdownIcon'));
+  fireEvent.click(getByTestId('clearChatButton'));
 
-  await waitFor(() => {});
+  fireEvent.click(getByTestId('ok-button'));
+
+  await waitFor(() => {
+    expect(setNotification).toHaveBeenCalled();
+  });
+});
+
+test('Contact: if not cache', async () => {
+  const chatMessages = (
+    <MemoryRouter>
+      <MockedProvider mocks={mocks} cache={cache}>
+        <ChatMessages entityId="3" />
+      </MockedProvider>
+    </MemoryRouter>
+  );
+  render(chatMessages);
+
+  await waitFor(() => {
+    expect(screen.getByTestId('beneficiaryName')).toHaveTextContent('New Contact');
+  });
 });
 
 const chatMessagesWithCollection = (
   <MemoryRouter>
-    <ApolloProvider client={client}>
+    <MockedProvider mocks={mocks} cache={cache}>
       <ChatMessages collectionId="2" />
-    </ApolloProvider>
+    </MockedProvider>
   </MemoryRouter>
 );
 
@@ -332,8 +537,6 @@ it('should have title as group name', async () => {
     expect(getByTestId('beneficiaryName')).toHaveTextContent('Default Group');
   });
 });
-
-// Need to scroll up first
 
 test('Collection: click on Jump to latest', async () => {
   const { getByTestId } = render(chatMessagesWithCollection);
@@ -368,121 +571,46 @@ test('should send message to contact collections', async () => {
   });
 });
 
+test('Collection: if cache', async () => {
+  cache.writeQuery(collection);
+
+  const chatMessagesWithCollection = (
+    <Router>
+      <MockedProvider cache={cache} mocks={mocks}>
+        <ChatMessages collectionId="2" />
+      </MockedProvider>
+    </Router>
+  );
+  render(chatMessagesWithCollection);
+
+  await waitFor(() => {
+    expect(screen.getByTestId('beneficiaryName')).toHaveTextContent('Default Group');
+  });
+});
+
 test('Collection: if not cache', async () => {
   const chatMessagesWithCollection = (
     <Router>
-      <MockedProvider mocks={[...CONVERSATION_MOCKS, ...mocksWithConversation]}>
+      <MockedProvider cache={cache} mocks={mocks}>
         <ChatMessages collectionId="5" />
       </MockedProvider>
     </Router>
   );
-  const { getByTestId } = render(chatMessagesWithCollection);
-  // need to check why we click this
-
-  await waitFor(() => {});
-});
-
-test('Collection: if cache', async () => {
-  cache.writeQuery(collection);
-
-  const client = new ApolloClient({
-    cache: cache,
-    uri: 'http://localhost:4000/',
-    assumeImmutableResults: true,
-  });
-  const chatMessagesWithCollection = (
-    <ApolloProvider client={client}>
-      <Router>
-        <MockedProvider mocks={[...CONVERSATION_MOCKS, ...mocksWithConversation]}>
-          <ChatMessages collectionId="5" />
-        </MockedProvider>
-      </Router>
-    </ApolloProvider>
-  );
   render(chatMessagesWithCollection);
 
-  // need to check why we click this
-
-  await waitFor(() => {});
-});
-
-test('click on Clear conversation', async () => {
-  const chatMessages = (
-    <MemoryRouter>
-      <ApolloProvider client={client}>
-        <ChatMessages entityId="2" />
-      </ApolloProvider>
-    </MemoryRouter>
-  );
-  const { getByTestId } = render(chatMessages);
   await waitFor(() => {
-    fireEvent.click(getByTestId('dropdownIcon'));
-    fireEvent.click(getByTestId('clearChatButton'));
-    // need to check this
-    // fireEvent.click(getByTestId('ok-button'));
-    // expect(getByTestId('app')).toHaveTextContent('Conversation cleared for this contact');
-  });
-});
-
-const messages = new Array(20).fill(body).map((b, index) => ({ ...b, id: `${index}` }));
-test.only('Load more messages', async () => {
-  const searchQuery = {
-    query: SEARCH_QUERY,
-    variables: {
-      filter: {},
-      contactOpts: { limit: DEFAULT_ENTITY_LIMIT },
-      messageOpts: { limit: DEFAULT_MESSAGE_LIMIT },
-    },
-    data: {
-      search: [
-        {
-          id: 'contact_2',
-          group: null,
-          contact: {
-            id: '2',
-            name: 'Effie Cormier',
-            phone: '987654321',
-            maskedPhone: '98****321',
-            lastMessageAt: '2020-06-29T09:31:47Z',
-            status: 'VALID',
-            fields: '{}',
-            bspStatus: 'SESSION_AND_HSM',
-            isOrgRead: true,
-          },
-          messages,
-        },
-      ],
-    },
-  };
-
-  cache.writeQuery(searchQuery);
-  const client = new ApolloClient({
-    cache: cache,
-    uri: 'http://localhost:4000/',
-    assumeImmutableResults: true,
-  });
-
-  const chatMessages = (
-    <MemoryRouter>
-      <MockedProvider mocks={[loadMoreChats]}>
-        <ApolloProvider client={client}>
-          <ChatMessages entityId="2" />
-        </ApolloProvider>
-      </MockedProvider>
-    </MemoryRouter>
-  );
-
-  const { getByTestId } = render(chatMessages);
-
-  await waitFor(() => {
-    const container: any = document.querySelector('.messageContainer');
-    fireEvent.scroll(container, { target: { scrollY: 0 } });
-    fireEvent.click(getByTestId('loadMoreMessages'));
+    expect(screen.getByTestId('beneficiaryName')).toHaveTextContent('Test 5');
   });
 });
 
 test('Should render for multi-search', async () => {
-  defineUrl('http://localhost:3000/chat/2?search=48');
+  const chatMessages = (
+    <MemoryRouter initialEntries={['/chat/2?search=48']}>
+      <MockedProvider mocks={mocks} cache={cache}>
+        <ChatMessages entityId="2" />
+      </MockedProvider>
+    </MemoryRouter>
+  );
 
   const { getByTestId } = render(chatMessages);
 
@@ -512,25 +640,30 @@ test('send message to contact', async () => {
   });
 });
 
+test('Load more messages', async () => {
+  const { getByTestId } = render(chatMessages);
+
+  await waitFor(() => {
+    const container: any = document.querySelector('.messageContainer');
+    fireEvent.scroll(container, { target: { scrollY: 0 } });
+    fireEvent.click(getByTestId('loadMoreMessages'));
+  });
+});
+
 const groupscache = new InMemoryCache({ addTypename: false });
 groupscache.writeQuery(waGroup);
 groupscache.writeQuery(waGroupcollection);
 
-const groupClient = new ApolloClient({
-  cache: groupscache,
-  uri: 'http://localhost:4000/',
-  assumeImmutableResults: true,
-});
-let route = '/group/chat';
+const route = '/group/chat';
 
 const chatMessagesWAGroups = (
-  <MemoryRouter initialEntries={[route]}>
-    <ApolloProvider client={groupClient}>
+  <MockedProvider cache={groupscache} mocks={mocks}>
+    <MemoryRouter initialEntries={[route]}>
       <Routes>
         <Route path="group/chat" element={<ChatMessages entityId="2" />} />
       </Routes>
-    </ApolloProvider>
-  </MemoryRouter>
+    </MemoryRouter>
+  </MockedProvider>
 );
 
 it('should have title as contact name for whatsapp groups', async () => {
@@ -561,11 +694,11 @@ test('send message to whatsapp group', async () => {
 
 const collectionMessagesWAGroups = (
   <MemoryRouter initialEntries={[route]}>
-    <ApolloProvider client={groupClient}>
+    <MockedProvider cache={groupscache} mocks={mocks}>
       <Routes>
         <Route path="group/chat" element={<ChatMessages collectionId="1" />} />
       </Routes>
-    </ApolloProvider>
+    </MockedProvider>
   </MemoryRouter>
 );
 
@@ -589,24 +722,16 @@ test('should send message to whatsapp group collections', async () => {
 });
 
 test('should show error if send message fails', async () => {
-  const { getByTestId } = render(
-    <MemoryRouter>
-      <MockedProvider mocks={[sendMessageMock]}>
-        <ApolloProvider client={client}>
-          <ChatMessages entityId="2" />
-        </ApolloProvider>
-      </MockedProvider>
-    </MemoryRouter>
-  );
+  const { getByTestId } = render(chatMessagesWAGroups);
 
   const editor = screen.getByTestId('editor');
 
   await userEvent.click(editor);
   await userEvent.tab();
-  fireEvent.input(editor, { data: 'test' });
+  fireEvent.input(editor, { data: 'hey' });
 
   await waitFor(() => {
-    expect(editor).toHaveTextContent('test');
+    expect(editor).toHaveTextContent('hey');
   });
 
   fireEvent.click(getByTestId('sendButton'), { force: true });
